ugc/ugc/src/requirements.txt
annotated-types==0.6.0
attrs==23.2.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
confluent-kafka==2.3.0
ecdsa==0.18.0
flasgger==0.9.7.1
flask==3.0.2
gevent==24.2.1
greenlet==3.0.3
gunicorn[gevent]==21.2.0
itsdangerous==2.1.2
jinja2==3.1.3
jsonschema-specifications==2023.12.1
jsonschema==4.21.1
markupsafe==2.1.5
mistune==3.0.2
packaging==24.0
pika-stubs==0.1.3
pika==1.3.2
pyasn1==0.6.0
pycparser==2.22
pydantic-core==2.16.3
pydantic-settings==2.2.1
pydantic==2.6.4
python-dotenv==1.0.1
python-jose==3.3.0
python-logstash==0.4.8
pyyaml==6.0.1
referencing==0.34.0
rpds-py==0.18.0
rsa==4.9
setuptools==69.2.0
sentry-sdk[flask]==1.45.0
six==1.16.0
types-gevent==24.2.0.20240315
types-greenlet==3.0.0.20240311
types-psutil==5.9.5.20240316
types-pyasn1==0.6.0.20240402
types-python-jose==3.3.4.20240106
typing-extensions==4.11.0
werkzeug==3.0.2
win32-setctime==1.1.0
zope-event==5.0
zope-interface==6.2
pymongo